refactor: move the parquet trace-repo plumbing into traces
resolve_parquet, download_shard and their helpers resolve a Hub trace dataset's
refs/convert/parquet branch and pull its shards. They sat in memory/hf_dataset.rs, but
they have nothing to do with a memory: their only caller is traces/source.rs, feeding the
parser in traces/parquet.rs. That was the crate's last real upward edge — traces reaching
into memory — and it's now traces → hub instead, which is a transport call like any other.
